+++ src/screens/Jobs/Ext_JobsInProgress.jsx
@@ -4,6 +4,7 @@ import JobsInProgress_Table from '../../components/JobsInProgress_Table';
 function JobsInProgress() {
   return (
     <div>
+      
       <JobsInProgress_Table/>
     </div>
   )
