+++ src/screens/Jobs/Ext_NewJobRequests.jsx
@@ -1,12 +1,13 @@
 import { Form, Input, InputNumber, Popconfirm, Table, Typography } from 'antd';
 import axios from 'axios';
 import React, { useState ,useEffect} from 'react';
+import JobsInProgress_Table from '../../components/JobsInProgress_Table'; 
 // import "../components/Job.css"
 import {useNavigate} from "react-router-dom"
 const originData = [];
 
 
-const CompletedJobs = () => {
+const NewJobRequests = () => {
   const [form] = Form.useForm();
   const [data, setData] = useState(originData);
   const [editingKey, setEditingKey] = useState('');
@@ -35,8 +36,7 @@ const CompletedJobs = () => {
             "Deliverer ID":e.deliverer_id,
             "Customer ID":e.customer_id,
             "Document ID":e.customer_id,
-            "Created date":e.createdAt,
-            "Completed date":e.updatedAt,
+            "Created date":e.createdAt
           }
         );
       })
@@ -85,13 +85,18 @@ const CompletedJobs = () => {
         width: '15%',
         editable: false,
       },
-      {
-        title: 'Completed date',
-        dataIndex: 'updatedAt',
-        width: '15%',
-        editable: false,
+    {
+      
+      // dataIndex: 'Profile',
+      render: (_, record) => {
+        const editable = isEditing(record);
+        return (
+          <Typography.Link disabled={editingKey !== ''} >
+            Accept
+          </Typography.Link>
+        );
       },
-
+    },
   ];
   const mergedColumns = columns.map((col) => {
     return {
@@ -102,9 +107,9 @@ const CompletedJobs = () => {
   return (
     <div className='Table'>
       
-        <h2>Completed Jobs</h2>
+        <h2>New Job Requests</h2>
         <Form form={form} component={false}>
-      <Table
+      < JobsInProgress_Table
         bordered
         dataSource={data}
         columns={mergedColumns}
@@ -115,5 +120,4 @@ const CompletedJobs = () => {
   )
 }
 
-export default CompletedJobs;
-
+export default NewJobRequests;
